+++ trl/models/__init__.py
@@ -26,7 +26,7 @@ from ..import_utils import is_diffusers_available
 
 
 if is_diffusers_available():
-    from .modelling_sd_base import (
+    from .modeling_sd_base import (
         DDPOPipelineOutput,
         DDPOSchedulerOutput,
         DDPOStableDiffusionPipeline,
